split canonical json redaction
